+++ android/301Project/app/src/main/java/com/example/a301project/IngredientController.java
@@ -75,7 +75,7 @@ public class IngredientController {
      */
     public void addIngredient(Ingredient ingredient) {
         // get all required values
-        int amount = ingredient.getAmount();
+        double amount = ingredient.getAmount();
         String bestBefore = ingredient.getbbd();
         String category = ingredient.getCategory();
         String location = ingredient.getLocation();
